test: factor out sharable utils
